修改了图片格式，请求方式等
