+++ frontend/nuxt.config.ts
@@ -1,4 +1,9 @@
 export default defineNuxtConfig({
   compatibilityDate: "2025-03-15",
   css: ["@fortawesome/fontawesome-svg-core/styles.css"],
+  vite: {
+    build: {
+      target: "es2015",
+    },
+  },
 });
